decentralized/requirements.txt
ipython==8.5.0
ipywidgets==8.0.2
joblib==1.1.0
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.3
prettytable==3.4.1
pytest==7.1.3
PyYAML==6.0
schema==0.7.5
scikit_learn==1.1.2
tqdm==4.64.1
